sdk: redesign sync events APIs
Ref https://github.com/rust-nostr/nostr/issues/920

Pull-Request: https://github.com/rust-nostr/nostr/pull/1235
