+++ crates/nostr-relay-builder/src/local/mod.rs
@@ -7,6 +7,7 @@
 use std::net::SocketAddr;
 
 use atomic_destructor::AtomicDestructor;
+use nostr_sdk::client::SyncSummary;
 use nostr_sdk::prelude::*;
 use tokio::io::{AsyncRead, AsyncWrite};
 
@@ -82,8 +83,8 @@ impl LocalRelay {
         &self,
         urls: I,
         filter: Filter,
-        opts: &SyncOptions,
-    ) -> Result<Output<Reconciliation>, Error>
+        opts: SyncOptions,
+    ) -> Result<Output<SyncSummary>, Error>
     where
         I: IntoIterator<Item = U>,
         U: Into<RelayUrlArg<'a>>,
