+++ sdk/src/relay/api/mod.rs
@@ -1,9 +1,11 @@
 mod fetch_events;
 mod stream_events;
 mod subscribe;
+mod sync;
 mod try_connect;
 
 pub use self::fetch_events::*;
 pub use self::stream_events::*;
 pub use self::subscribe::*;
+pub use self::sync::*;
 pub use self::try_connect::*;
