+++ sdk/src/relay/inner.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 use std::cmp;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
@@ -8,7 +8,6 @@ use std::time::Duration;
 use async_utility::{task, time};
 use async_wsocket::futures_util::{self, SinkExt, StreamExt};
 use async_wsocket::{ConnectionMode, Message};
-use negentropy::{Id, Negentropy, NegentropyStorageVector};
 use nostr::rand::rngs::OsRng;
 use nostr::rand::{Rng, RngCore, TryRngCore};
 use nostr_database::prelude::*;
@@ -18,16 +17,13 @@ use tokio::sync::{broadcast, Mutex, MutexGuard, Notify, RwLock, RwLockWriteGuard
 use super::capabilities::{AtomicRelayCapabilities, RelayCapabilities};
 use super::constants::{
     DEFAULT_CONNECTION_TIMEOUT, JITTER_RANGE, MAX_RETRY_INTERVAL, MIN_ATTEMPTS, MIN_SUCCESS_RATE,
-    NEGENTROPY_BATCH_SIZE_DOWN, NEGENTROPY_FRAME_SIZE_LIMIT, NEGENTROPY_HIGH_WATER_UP,
-    NEGENTROPY_LOW_WATER_UP, PING_INTERVAL, SLEEP_INTERVAL, WAIT_FOR_OK_TIMEOUT,
-    WEBSOCKET_TX_TIMEOUT,
+    PING_INTERVAL, SLEEP_INTERVAL, WAIT_FOR_OK_TIMEOUT, WEBSOCKET_TX_TIMEOUT,
 };
-use super::options::{RelayOptions, ReqExitPolicy, SubscribeAutoCloseOptions, SyncOptions};
+use super::options::{RelayOptions, ReqExitPolicy, SubscribeAutoCloseOptions};
 use super::ping::PingTracker;
 use super::stats::RelayConnectionStats;
 use super::{
-    Error, Reconciliation, RelayNotification, RelayStatus, SubscriptionActivity,
-    SubscriptionAutoClosedReason,
+    Error, RelayNotification, RelayStatus, SubscriptionActivity, SubscriptionAutoClosedReason,
 };
 use crate::policy::AdmitStatus;
 use crate::pool::RelayPoolNotification;
@@ -1328,19 +1324,6 @@ impl InnerRelay {
         self.atomic.channels.send_client_msgs(msgs)
     }
 
-    fn send_neg_msg(&self, id: &SubscriptionId, message: &str) -> Result<(), Error> {
-        self.send_msg(ClientMessage::NegMsg {
-            subscription_id: Cow::Borrowed(id),
-            message: Cow::Borrowed(message),
-        })
-    }
-
-    fn send_neg_close(&self, id: &SubscriptionId) -> Result<(), Error> {
-        self.send_msg(ClientMessage::NegClose {
-            subscription_id: Cow::Borrowed(id),
-        })
-    }
-
     async fn auth(&self, challenge: String) -> Result<(), Error> {
         // Get signer
         let signer = self.state.signer().await?;
@@ -1728,406 +1711,6 @@ impl InnerRelay {
 
         Ok(())
     }
-
-    #[inline(never)]
-    fn handle_neg_msg<I>(
-        &self,
-        subscription_id: &SubscriptionId,
-        msg: Option<Vec<u8>>,
-        curr_have_ids: I,
-        curr_need_ids: I,
-        opts: &SyncOptions,
-        output: &mut Reconciliation,
-        have_ids: &mut Vec<EventId>,
-        need_ids: &mut Vec<EventId>,
-        sync_done: &mut bool,
-    ) -> Result<(), Error>
-    where
-        I: Iterator<Item = EventId>,
-    {
-        let mut counter: u64 = 0;
-
-        // If event ID wasn't already seen, add to the HAVE IDs
-        // Add to HAVE IDs only if `do_up` is true
-        for id in curr_have_ids.into_iter() {
-            if output.local.insert(id) && opts.do_up() {
-                have_ids.push(id);
-                counter += 1;
-            }
-        }
-
-        // If event ID wasn't already seen, add to the NEED IDs
-        // Add to NEED IDs only if `do_down` is true
-        for id in curr_need_ids.into_iter() {
-            if output.remote.insert(id) && opts.do_down() {
-                need_ids.push(id);
-                counter += 1;
-            }
-        }
-
-        if let Some(progress) = &opts.progress {
-            progress.send_modify(|state| {
-                state.total += counter;
-            });
-        }
-
-        match msg {
-            Some(query) => self.send_neg_msg(subscription_id, &hex::encode(query)),
-            None => {
-                // Mark sync as done
-                *sync_done = true;
-
-                // Send NEG-CLOSE message
-                self.send_neg_close(subscription_id)
-            }
-        }
-    }
-
-    #[inline(never)]
-    async fn upload_neg_events(
-        &self,
-        have_ids: &mut Vec<EventId>,
-        in_flight_up: &mut HashSet<EventId>,
-        opts: &SyncOptions,
-    ) -> Result<(), Error> {
-        // Check if it should skip the upload
-        if !opts.do_up() || have_ids.is_empty() || in_flight_up.len() > NEGENTROPY_LOW_WATER_UP {
-            return Ok(());
-        }
-
-        let mut num_sent = 0;
-
-        while !have_ids.is_empty() && in_flight_up.len() < NEGENTROPY_HIGH_WATER_UP {
-            if let Some(id) = have_ids.pop() {
-                match self.state.database().event_by_id(&id).await {
-                    Ok(Some(event)) => {
-                        in_flight_up.insert(id);
-                        self.send_msg(ClientMessage::event(event))?;
-                        num_sent += 1;
-                    }
-                    Ok(None) => {
-                        // Event not found
-                    }
-                    Err(e) => tracing::error!(
-                        url = %self.url,
-                        error = %e,
-                        "Can't upload event."
-                    ),
-                }
-            }
-        }
-
-        // Update progress
-        if let Some(progress) = &opts.progress {
-            progress.send_modify(|state| {
-                state.current += num_sent;
-            });
-        }
-
-        if num_sent > 0 {
-            tracing::info!(
-                "Negentropy UP for '{}': {} events ({} remaining)",
-                self.url,
-                num_sent,
-                have_ids.len()
-            );
-        }
-
-        Ok(())
-    }
-
-    #[inline(never)]
-    async fn req_neg_events(
-        &self,
-        need_ids: &mut Vec<EventId>,
-        in_flight_down: &mut bool,
-        down_sub_id: &SubscriptionId,
-        opts: &SyncOptions,
-    ) -> Result<(), Error> {
-        // Check if it should skip the download
-        if !opts.do_down() || need_ids.is_empty() || *in_flight_down {
-            return Ok(());
-        }
-
-        let capacity: usize = cmp::min(need_ids.len(), NEGENTROPY_BATCH_SIZE_DOWN);
-        let mut ids: Vec<EventId> = Vec::with_capacity(capacity);
-
-        while !need_ids.is_empty() && ids.len() < NEGENTROPY_BATCH_SIZE_DOWN {
-            if let Some(id) = need_ids.pop() {
-                ids.push(id);
-            }
-        }
-
-        tracing::info!(
-            "Negentropy DOWN for '{}': {} events ({} remaining)",
-            self.url,
-            ids.len(),
-            need_ids.len()
-        );
-
-        // Update progress
-        if let Some(progress) = &opts.progress {
-            progress.send_modify(|state| {
-                state.current += ids.len() as u64;
-            });
-        }
-
-        let filter = Filter::new().ids(ids);
-        let msg: ClientMessage = ClientMessage::Req {
-            subscription_id: Cow::Borrowed(down_sub_id),
-            filters: vec![Cow::Borrowed(&filter)],
-        };
-
-        // Register an auto-closing subscription
-        self.add_auto_closing_subscription(down_sub_id.clone(), vec![filter.clone()])
-            .await;
-
-        // Send msg
-        if let Err(e) = self.send_msg(msg) {
-            // Remove previously added subscription
-            self.remove_subscription(down_sub_id).await;
-
-            // Propagate error
-            return Err(e);
-        }
-
-        *in_flight_down = true;
-
-        Ok(())
-    }
-
-    /// Returns `true` if the events was in the `in_flight_up` collection.
-    #[inline(never)]
-    fn handle_neg_ok(
-        &self,
-        in_flight_up: &mut HashSet<EventId>,
-        event_id: EventId,
-        status: bool,
-        message: Cow<'_, str>,
-        output: &mut Reconciliation,
-    ) -> bool {
-        if in_flight_up.remove(&event_id) {
-            if status {
-                output.sent.insert(event_id);
-            } else {
-                tracing::error!(
-                    url = %self.url,
-                    id = %event_id,
-                    msg = %message,
-                    "Can't upload event."
-                );
-
-                output
-                    .send_failures
-                    .entry(self.url.clone())
-                    .and_modify(|map| {
-                        map.insert(event_id, message.to_string());
-                    })
-                    .or_default()
-                    .insert(event_id, message.into_owned());
-            }
-
-            true
-        } else {
-            false
-        }
-    }
-
-    /// New negentropy protocol
-    #[inline(never)]
-    pub(super) async fn sync(
-        &self,
-        filter: &Filter,
-        items: Vec<(EventId, Timestamp)>,
-        opts: &SyncOptions,
-        output: &mut Reconciliation,
-    ) -> Result<(), Error> {
-        // Prepare the negentropy client
-        let storage: NegentropyStorageVector = prepare_negentropy_storage(items)?;
-        let mut negentropy: Negentropy<NegentropyStorageVector> =
-            Negentropy::borrowed(&storage, NEGENTROPY_FRAME_SIZE_LIMIT)?;
-
-        // Initiate reconciliation
-        let initial_message: Vec<u8> = negentropy.initiate()?;
-
-        // Subscribe
-        let mut notifications = self.internal_notification_sender.subscribe();
-        let mut temp_notifications = self.internal_notification_sender.subscribe();
-
-        // Send the initial negentropy message
-        let sub_id: SubscriptionId = SubscriptionId::generate();
-        let open_msg: ClientMessage = ClientMessage::NegOpen {
-            subscription_id: Cow::Borrowed(&sub_id),
-            filter: Cow::Borrowed(filter),
-            id_size: None,
-            initial_message: Cow::Owned(hex::encode(initial_message)),
-        };
-        self.send_msg(open_msg)?;
-
-        // Check if negentropy is supported
-        check_negentropy_support(&sub_id, opts, &mut temp_notifications).await?;
-
-        let mut in_flight_up: HashSet<EventId> = HashSet::new();
-        let mut in_flight_down: bool = false;
-        let mut sync_done: bool = false;
-        let mut have_ids: Vec<EventId> = Vec::new();
-        let mut need_ids: Vec<EventId> = Vec::new();
-        let down_sub_id: SubscriptionId = SubscriptionId::generate();
-        let mut last_relevant_msg: Instant = Instant::now();
-
-        // Start reconciliation
-        while let Ok(notification) = notifications.recv().await {
-            if last_relevant_msg.elapsed() > opts.idle_timeout {
-                return Err(Error::Timeout);
-            }
-
-            match notification {
-                RelayNotification::Message { message } => {
-                    let is_relevant: bool = match message {
-                        RelayMessage::NegMsg {
-                            subscription_id,
-                            message,
-                        } => {
-                            if subscription_id.as_ref() == &sub_id {
-                                let mut curr_have_ids: Vec<Id> = Vec::new();
-                                let mut curr_need_ids: Vec<Id> = Vec::new();
-
-                                // Parse message
-                                let query: Vec<u8> = hex::decode(message.as_ref())?;
-
-                                // Reconcile
-                                let msg: Option<Vec<u8>> = negentropy.reconcile_with_ids(
-                                    &query,
-                                    &mut curr_have_ids,
-                                    &mut curr_need_ids,
-                                )?;
-
-                                // Handle the message
-                                self.handle_neg_msg(
-                                    &subscription_id,
-                                    msg,
-                                    curr_have_ids.into_iter().map(neg_id_to_event_id),
-                                    curr_need_ids.into_iter().map(neg_id_to_event_id),
-                                    opts,
-                                    output,
-                                    &mut have_ids,
-                                    &mut need_ids,
-                                    &mut sync_done,
-                                )?;
-
-                                // Relevant to this sync
-                                true
-                            } else {
-                                // Not relevant to this sync
-                                false
-                            }
-                        }
-                        RelayMessage::NegErr {
-                            subscription_id,
-                            message,
-                        } => {
-                            if subscription_id.as_ref() == &sub_id {
-                                return Err(Error::RelayMessage(message.into_owned()));
-                            } else {
-                                // Not relevant to this sync
-                                false
-                            }
-                        }
-                        RelayMessage::Ok {
-                            event_id,
-                            status,
-                            message,
-                        } => {
-                            self.handle_neg_ok(&mut in_flight_up, event_id, status, message, output)
-                        }
-                        RelayMessage::Event {
-                            subscription_id,
-                            event,
-                        } => {
-                            if subscription_id.as_ref() == &down_sub_id {
-                                output.received.insert(event.id);
-
-                                // Relevant to this sync
-                                true
-                            } else {
-                                // Not relevant to this sync
-                                false
-                            }
-                        }
-                        RelayMessage::EndOfStoredEvents(subscription_id) => {
-                            if subscription_id.as_ref() == &down_sub_id {
-                                in_flight_down = false;
-
-                                // Remove subscription
-                                self.remove_subscription(&down_sub_id).await;
-
-                                // Close subscription
-                                self.send_msg(ClientMessage::Close(Cow::Borrowed(&down_sub_id)))?;
-
-                                // Relevant to this sync
-                                true
-                            } else {
-                                // Not relevant to this sync
-                                false
-                            }
-                        }
-                        RelayMessage::Closed {
-                            subscription_id, ..
-                        } => {
-                            if subscription_id.as_ref() == &down_sub_id {
-                                in_flight_down = false;
-
-                                // NOTE: the subscription is removed in the `InnerRelay::handle_relay_message` method,
-                                // so there is no need to try to remove it also here.
-
-                                // Relevant to this sync
-                                true
-                            } else {
-                                // Not relevant to this sync
-                                false
-                            }
-                        }
-                        _ => false,
-                    };
-
-                    // Send events
-                    self.upload_neg_events(&mut have_ids, &mut in_flight_up, opts)
-                        .await?;
-
-                    // Get events
-                    self.req_neg_events(&mut need_ids, &mut in_flight_down, &down_sub_id, opts)
-                        .await?;
-
-                    // NOTE: update this after the uploading and requesting of the events, as it may require some time.
-                    if is_relevant {
-                        last_relevant_msg = Instant::now();
-                    }
-                }
-                RelayNotification::RelayStatus { status } => {
-                    if status.is_disconnected() {
-                        return Err(Error::NotConnected);
-                    }
-                }
-                RelayNotification::Shutdown => {
-                    return Err(Error::ReceivedShutdown);
-                }
-                _ => (),
-            };
-
-            if sync_done
-                && have_ids.is_empty()
-                && need_ids.is_empty()
-                && in_flight_up.is_empty()
-                && !in_flight_down
-            {
-                break;
-            }
-        }
-
-        tracing::info!(url = %self.url, "Negentropy reconciliation terminated.");
-
-        Ok(())
-    }
 }
 
 /// Send WebSocket messages with timeout set to [WEBSOCKET_TX_TIMEOUT].
@@ -2146,87 +1729,6 @@ async fn close_ws(tx: &mut WebSocketSink) -> Result<(), Error> {
         Some(res) => Ok(res?),
         None => Err(Error::Timeout),
     }
-}
-
-#[inline]
-fn neg_id_to_event_id(id: Id) -> EventId {
-    EventId::from_byte_array(id.to_bytes())
-}
-
-fn prepare_negentropy_storage(
-    items: Vec<(EventId, Timestamp)>,
-) -> Result<NegentropyStorageVector, Error> {
-    // Compose negentropy storage
-    let mut storage = NegentropyStorageVector::with_capacity(items.len());
-
-    // Add items
-    for (id, timestamp) in items.into_iter() {
-        let id: Id = Id::from_byte_array(id.to_bytes());
-        storage.insert(timestamp.as_secs(), id)?;
-    }
-
-    // Seal
-    storage.seal()?;
-
-    // Build negentropy client
-    Ok(storage)
-}
-
-/// Check if negentropy is supported
-#[inline(never)]
-async fn check_negentropy_support(
-    sub_id: &SubscriptionId,
-    opts: &SyncOptions,
-    temp_notifications: &mut broadcast::Receiver<RelayNotification>,
-) -> Result<(), Error> {
-    time::timeout(Some(opts.initial_timeout), async {
-        while let Ok(notification) = temp_notifications.recv().await {
-            if let RelayNotification::Message { message } = notification {
-                match message {
-                    RelayMessage::NegMsg {
-                        subscription_id, ..
-                    } => {
-                        if subscription_id.as_ref() == sub_id {
-                            break;
-                        }
-                    }
-                    RelayMessage::NegErr {
-                        subscription_id,
-                        message,
-                    } => {
-                        if subscription_id.as_ref() == sub_id {
-                            return Err(Error::RelayMessage(message.into_owned()));
-                        }
-                    }
-                    RelayMessage::Notice(message) => {
-                        if message == "ERROR: negentropy error: negentropy query missing elements" {
-                            // The NEG-OPEN message is sent with 4 elements instead of 5
-                            // If the relay return this error means that is not support new
-                            // negentropy protocol
-                            return Err(Error::Negentropy(
-                                negentropy::Error::UnsupportedProtocolVersion,
-                            ));
-                        } else if message.contains("bad msg")
-                            && (message.contains("unknown cmd")
-                                || message.contains("negentropy")
-                                || message.contains("NEG-"))
-                        {
-                            return Err(Error::NegentropyNotSupported);
-                        } else if message.contains("bad msg: invalid message")
-                            && message.contains("NEG-OPEN")
-                        {
-                            return Err(Error::UnknownNegentropyError);
-                        }
-                    }
-                    _ => (),
-                }
-            }
-        }
-
-        Ok(())
-    })
-    .await
-    .ok_or(Error::Timeout)?
 }
 
 #[cfg(bench)]
